Extract nine patch segment generation into create_nine_patch_segments.
