+++ webrender/src/border.rs
@@ -5,8 +5,9 @@
 use api::{BorderRadius, BorderSide, BorderStyle, ColorF, ColorU, DeviceRect, DeviceSize};
 use api::{LayoutSideOffsets, LayoutSizeAu, LayoutPrimitiveInfo, LayoutToDeviceScale};
 use api::{DeviceVector2D, DevicePoint, LayoutRect, LayoutSize, NormalBorder, DeviceIntSize};
-use api::{AuHelpers};
+use api::{AuHelpers, NinePatchBorder, LayoutPoint, RepeatMode, TexelRect};
 use ellipse::Ellipse;
+use euclid::vec2;
 use display_list_flattener::DisplayListFlattener;
 use gpu_types::{BorderInstance, BorderSegment, BrushFlags};
 use prim_store::{BorderSegmentInfo, BrushKind, BrushPrimitive, BrushSegment, BrushSegmentVec};
@@ -1140,4 +1141,173 @@ pub fn create_normal_border_prim(
             segments: brush_segments,
         }),
     )
+}
+
+pub fn create_nine_patch_segments(
+    rect: &LayoutRect,
+    widths: &LayoutSideOffsets,
+    border: &NinePatchBorder,
+) -> BrushSegmentDescriptor {
+    // Calculate the modified rect as specific by border-image-outset
+    let origin = LayoutPoint::new(
+        rect.origin.x - border.outset.left,
+        rect.origin.y - border.outset.top,
+    );
+    let size = LayoutSize::new(
+        rect.size.width + border.outset.left + border.outset.right,
+        rect.size.height + border.outset.top + border.outset.bottom,
+    );
+    let rect = LayoutRect::new(origin, size);
+
+    // Calculate the local texel coords of the slices.
+    let px0 = 0.0;
+    let px1 = border.slice.left as f32;
+    let px2 = border.width as f32 - border.slice.right as f32;
+    let px3 = border.width as f32;
+
+    let py0 = 0.0;
+    let py1 = border.slice.top as f32;
+    let py2 = border.height as f32 - border.slice.bottom as f32;
+    let py3 = border.height as f32;
+
+    let tl_outer = LayoutPoint::new(rect.origin.x, rect.origin.y);
+    let tl_inner = tl_outer + vec2(widths.left, widths.top);
+
+    let tr_outer = LayoutPoint::new(rect.origin.x + rect.size.width, rect.origin.y);
+    let tr_inner = tr_outer + vec2(-widths.right, widths.top);
+
+    let bl_outer = LayoutPoint::new(rect.origin.x, rect.origin.y + rect.size.height);
+    let bl_inner = bl_outer + vec2(widths.left, -widths.bottom);
+
+    let br_outer = LayoutPoint::new(
+        rect.origin.x + rect.size.width,
+        rect.origin.y + rect.size.height,
+    );
+    let br_inner = br_outer - vec2(widths.right, widths.bottom);
+
+    fn add_segment(
+        segments: &mut BrushSegmentVec,
+        rect: LayoutRect,
+        uv_rect: TexelRect,
+        repeat_horizontal: RepeatMode,
+        repeat_vertical: RepeatMode
+    ) {
+        if uv_rect.uv1.x > uv_rect.uv0.x &&
+           uv_rect.uv1.y > uv_rect.uv0.y {
+
+            // Use segment relative interpolation for all
+            // instances in this primitive.
+            let mut brush_flags =
+                BrushFlags::SEGMENT_RELATIVE |
+                BrushFlags::SEGMENT_TEXEL_RECT;
+
+            // Enable repeat modes on the segment.
+            if repeat_horizontal == RepeatMode::Repeat {
+                brush_flags |= BrushFlags::SEGMENT_REPEAT_X;
+            }
+            if repeat_vertical == RepeatMode::Repeat {
+                brush_flags |= BrushFlags::SEGMENT_REPEAT_Y;
+            }
+
+            let segment = BrushSegment::new(
+                rect,
+                true,
+                EdgeAaSegmentMask::empty(),
+                [
+                    uv_rect.uv0.x,
+                    uv_rect.uv0.y,
+                    uv_rect.uv1.x,
+                    uv_rect.uv1.y,
+                ],
+                brush_flags,
+            );
+
+            segments.push(segment);
+        }
+    }
+
+    // Build the list of image segments
+    let mut segments = BrushSegmentVec::new();
+
+    // Top left
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(tl_outer.x, tl_outer.y, tl_inner.x, tl_inner.y),
+        TexelRect::new(px0, py0, px1, py1),
+        RepeatMode::Stretch,
+        RepeatMode::Stretch
+    );
+    // Top right
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(tr_inner.x, tr_outer.y, tr_outer.x, tr_inner.y),
+        TexelRect::new(px2, py0, px3, py1),
+        RepeatMode::Stretch,
+        RepeatMode::Stretch
+    );
+    // Bottom right
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(br_inner.x, br_inner.y, br_outer.x, br_outer.y),
+        TexelRect::new(px2, py2, px3, py3),
+        RepeatMode::Stretch,
+        RepeatMode::Stretch
+    );
+    // Bottom left
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(bl_outer.x, bl_inner.y, bl_inner.x, bl_outer.y),
+        TexelRect::new(px0, py2, px1, py3),
+        RepeatMode::Stretch,
+        RepeatMode::Stretch
+    );
+
+    // Center
+    if border.fill {
+        add_segment(
+            &mut segments,
+            LayoutRect::from_floats(tl_inner.x, tl_inner.y, tr_inner.x, bl_inner.y),
+            TexelRect::new(px1, py1, px2, py2),
+            border.repeat_horizontal,
+            border.repeat_vertical
+        );
+    }
+
+    // Add edge segments.
+
+    // Top
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(tl_inner.x, tl_outer.y, tr_inner.x, tl_inner.y),
+        TexelRect::new(px1, py0, px2, py1),
+        border.repeat_horizontal,
+        RepeatMode::Stretch,
+    );
+    // Bottom
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(bl_inner.x, bl_inner.y, br_inner.x, bl_outer.y),
+        TexelRect::new(px1, py2, px2, py3),
+        border.repeat_horizontal,
+        RepeatMode::Stretch,
+    );
+    // Left
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(tl_outer.x, tl_inner.y, tl_inner.x, bl_inner.y),
+        TexelRect::new(px0, py1, px1, py2),
+        RepeatMode::Stretch,
+        border.repeat_vertical,
+    );
+    // Right
+    add_segment(
+        &mut segments,
+        LayoutRect::from_floats(tr_inner.x, tr_inner.y, br_outer.x, br_inner.y),
+        TexelRect::new(px2, py1, px3, py2),
+        RepeatMode::Stretch,
+        border.repeat_vertical,
+    );
+    BrushSegmentDescriptor {
+        segments,
+    }
 }
