+++ webrender/src/display_list_flattener.rs
@@ -10,23 +10,22 @@ use api::{FilterOp, FontInstanceKey, GlyphInstance, GlyphOptions, RasterSpace, G
 use api::{IframeDisplayItem, ImageKey, ImageRendering, ItemRange, LayoutPoint, ColorDepth};
 use api::{LayoutPrimitiveInfo, LayoutRect, LayoutSize, LayoutTransform, LayoutVector2D};
 use api::{LineOrientation, LineStyle, NinePatchBorderSource, PipelineId};
-use api::{PropertyBinding, ReferenceFrame, RepeatMode, ScrollFrameDisplayItem, ScrollSensitivity};
+use api::{PropertyBinding, ReferenceFrame, ScrollFrameDisplayItem, ScrollSensitivity};
 use api::{Shadow, SpecificDisplayItem, StackingContext, StickyFrameDisplayItem, TexelRect};
 use api::{ClipMode, TransformStyle, YuvColorSpace, YuvData};
+use border::create_nine_patch_segments;
 use clip::{ClipChainId, ClipRegion, ClipItemKey, ClipStore, ClipItemSceneData};
 use clip_scroll_tree::{ROOT_SPATIAL_NODE_INDEX, ClipScrollTree, SpatialNodeIndex};
-use euclid::vec2;
 use frame_builder::{ChasePrimitive, FrameBuilder, FrameBuilderConfig};
 use glyph_rasterizer::FontInstance;
 use gpu_cache::GpuCacheHandle;
-use gpu_types::BrushFlags;
 use hit_test::{HitTestingItem, HitTestingRun};
 use image::simplify_repeated_primitive;
 use internal_types::{FastHashMap, FastHashSet};
 use picture::{Picture3DContext, PictureCompositeMode, PicturePrimitive, PrimitiveList};
-use prim_store::{BrushKind, BrushPrimitive, BrushSegmentDescriptor, PrimitiveInstance, PrimitiveDataInterner, PrimitiveKeyKind};
-use prim_store::{EdgeAaSegmentMask, ImageSource, PrimitiveOpacity, PrimitiveKey, PrimitiveSceneData, PrimitiveInstanceKind};
-use prim_store::{BorderSource, BrushSegment, BrushSegmentVec, PrimitiveContainer, PrimitiveDataHandle, PrimitiveStore};
+use prim_store::{BrushKind, BrushPrimitive, PrimitiveInstance, PrimitiveDataInterner, PrimitiveKeyKind};
+use prim_store::{ImageSource, PrimitiveOpacity, PrimitiveKey, PrimitiveSceneData, PrimitiveInstanceKind};
+use prim_store::{BorderSource, PrimitiveContainer, PrimitiveDataHandle, PrimitiveStore};
 use prim_store::{OpacityBinding, ScrollNodeAndClipChain, PictureIndex, register_prim_chase_id};
 use render_backend::{DocumentView};
 use resource_cache::{FontInstanceMap, ImageRequest};
@@ -36,7 +35,7 @@ use spatial_node::{StickyFrameInfo};
 use std::{f32, mem};
 use std::collections::vec_deque::VecDeque;
 use tiling::{CompositeOps};
-use util::{MaxRect, RectHelpers};
+use util::{MaxRect};
 
 #[derive(Debug, Copy, Clone)]
 struct ClipNode {
@@ -1728,171 +1727,13 @@ impl<'a> DisplayListFlattener<'a> {
         gradient_stops: ItemRange<GradientStop>,
         pipeline_id: PipelineId,
     ) {
-        let rect = info.rect;
         match border_item.details {
             BorderDetails::NinePatch(ref border) => {
-                // Calculate the modified rect as specific by border-image-outset
-                let origin = LayoutPoint::new(
-                    rect.origin.x - border.outset.left,
-                    rect.origin.y - border.outset.top,
+                let descriptor = create_nine_patch_segments(
+                    &info.rect,
+                    &border_item.widths,
+                    border,
                 );
-                let size = LayoutSize::new(
-                    rect.size.width + border.outset.left + border.outset.right,
-                    rect.size.height + border.outset.top + border.outset.bottom,
-                );
-                let rect = LayoutRect::new(origin, size);
-
-                // Calculate the local texel coords of the slices.
-                let px0 = 0.0;
-                let px1 = border.slice.left as f32;
-                let px2 = border.width as f32 - border.slice.right as f32;
-                let px3 = border.width as f32;
-
-                let py0 = 0.0;
-                let py1 = border.slice.top as f32;
-                let py2 = border.height as f32 - border.slice.bottom as f32;
-                let py3 = border.height as f32;
-
-                let tl_outer = LayoutPoint::new(rect.origin.x, rect.origin.y);
-                let tl_inner = tl_outer + vec2(border_item.widths.left, border_item.widths.top);
-
-                let tr_outer = LayoutPoint::new(rect.origin.x + rect.size.width, rect.origin.y);
-                let tr_inner = tr_outer + vec2(-border_item.widths.right, border_item.widths.top);
-
-                let bl_outer = LayoutPoint::new(rect.origin.x, rect.origin.y + rect.size.height);
-                let bl_inner = bl_outer + vec2(border_item.widths.left, -border_item.widths.bottom);
-
-                let br_outer = LayoutPoint::new(
-                    rect.origin.x + rect.size.width,
-                    rect.origin.y + rect.size.height,
-                );
-                let br_inner = br_outer - vec2(border_item.widths.right, border_item.widths.bottom);
-
-                fn add_segment(
-                    segments: &mut BrushSegmentVec,
-                    rect: LayoutRect,
-                    uv_rect: TexelRect,
-                    repeat_horizontal: RepeatMode,
-                    repeat_vertical: RepeatMode
-                ) {
-                    if uv_rect.uv1.x > uv_rect.uv0.x &&
-                       uv_rect.uv1.y > uv_rect.uv0.y {
-
-                        // Use segment relative interpolation for all
-                        // instances in this primitive.
-                        let mut brush_flags =
-                            BrushFlags::SEGMENT_RELATIVE |
-                            BrushFlags::SEGMENT_TEXEL_RECT;
-
-                        // Enable repeat modes on the segment.
-                        if repeat_horizontal == RepeatMode::Repeat {
-                            brush_flags |= BrushFlags::SEGMENT_REPEAT_X;
-                        }
-                        if repeat_vertical == RepeatMode::Repeat {
-                            brush_flags |= BrushFlags::SEGMENT_REPEAT_Y;
-                        }
-
-                        let segment = BrushSegment::new(
-                            rect,
-                            true,
-                            EdgeAaSegmentMask::empty(),
-                            [
-                                uv_rect.uv0.x,
-                                uv_rect.uv0.y,
-                                uv_rect.uv1.x,
-                                uv_rect.uv1.y,
-                            ],
-                            brush_flags,
-                        );
-
-                        segments.push(segment);
-                    }
-                }
-
-                // Build the list of image segments
-                let mut segments = BrushSegmentVec::new();
-
-                // Top left
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(tl_outer.x, tl_outer.y, tl_inner.x, tl_inner.y),
-                    TexelRect::new(px0, py0, px1, py1),
-                    RepeatMode::Stretch,
-                    RepeatMode::Stretch
-                );
-                // Top right
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(tr_inner.x, tr_outer.y, tr_outer.x, tr_inner.y),
-                    TexelRect::new(px2, py0, px3, py1),
-                    RepeatMode::Stretch,
-                    RepeatMode::Stretch
-                );
-                // Bottom right
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(br_inner.x, br_inner.y, br_outer.x, br_outer.y),
-                    TexelRect::new(px2, py2, px3, py3),
-                    RepeatMode::Stretch,
-                    RepeatMode::Stretch
-                );
-                // Bottom left
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(bl_outer.x, bl_inner.y, bl_inner.x, bl_outer.y),
-                    TexelRect::new(px0, py2, px1, py3),
-                    RepeatMode::Stretch,
-                    RepeatMode::Stretch
-                );
-
-                // Center
-                if border.fill {
-                    add_segment(
-                        &mut segments,
-                        LayoutRect::from_floats(tl_inner.x, tl_inner.y, tr_inner.x, bl_inner.y),
-                        TexelRect::new(px1, py1, px2, py2),
-                        border.repeat_horizontal,
-                        border.repeat_vertical
-                    );
-                }
-
-                // Add edge segments.
-
-                // Top
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(tl_inner.x, tl_outer.y, tr_inner.x, tl_inner.y),
-                    TexelRect::new(px1, py0, px2, py1),
-                    border.repeat_horizontal,
-                    RepeatMode::Stretch,
-                );
-                // Bottom
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(bl_inner.x, bl_inner.y, br_inner.x, bl_outer.y),
-                    TexelRect::new(px1, py2, px2, py3),
-                    border.repeat_horizontal,
-                    RepeatMode::Stretch,
-                );
-                // Left
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(tl_outer.x, tl_inner.y, tl_inner.x, bl_inner.y),
-                    TexelRect::new(px0, py1, px1, py2),
-                    RepeatMode::Stretch,
-                    border.repeat_vertical,
-                );
-                // Right
-                add_segment(
-                    &mut segments,
-                    LayoutRect::from_floats(tr_inner.x, tr_inner.y, br_outer.x, br_inner.y),
-                    TexelRect::new(px2, py1, px3, py2),
-                    RepeatMode::Stretch,
-                    border.repeat_vertical,
-                );
-                let descriptor = BrushSegmentDescriptor {
-                    segments,
-                };
 
                 let brush_kind = match border.source {
                     NinePatchBorderSource::Image(image_key) => {
